I won't lie: @feross' experience weighed heavy on me.
If such a fantastic open source engineer is struggling to make it work
financially, I don't think the ideal open source thing I had in mind for
Keykapp would. This is my life's work. It's all I want to do
for the rest of my life, and I have to find a way to make it work.